requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
boto3==1.17.95
botocore==1.20.95
certifi==2021.5.30
charset-normalizer==2.0.4
convertapi==1.4.0
dj-config-url==0.1.1
Django==3.2.4
django-storages==1.11.1
django-widget-tweaks==1.4.8
django-wkhtmltopdf==3.3.0
future==0.18.2
html5lib==1.1
idna==3.2
imgkit==1.2.2
jmespath==0.10.0
Pillow==8.2.0
psycopg2-binary==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.68
requests==2.26.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
webencodings==0.5.1
xhtml2pdf==0.2.5
